Rename Observable to EventProducer. Implement new Observable as the EventProducer with a value.
