+++ Bond/Extensions/iOS/UICollectionView+Bond.swift
@@ -153,9 +153,9 @@ extension UICollectionView {
   }
 }
 
-public extension ObservableType where
+public extension EventProducerType where
   EventType: ObservableArrayEventType,
-  EventType.ObservableArrayEventSequenceType.Generator.Element: ObservableType,
+  EventType.ObservableArrayEventSequenceType.Generator.Element: EventProducerType,
   EventType.ObservableArrayEventSequenceType.Generator.Element.EventType: ObservableArrayEventType {
   
   private typealias ElementType = EventType.ObservableArrayEventSequenceType.Generator.Element.EventType.ObservableArrayEventSequenceType.Generator.Element
